mt5-websocket-client/requirements.txt
websockets==10.4
python-dotenv==1.0.0
google-cloud-storage==2.9.0